pyproject.toml
[tool.poetry]
name = "shop"
version = "0.1.0"
description = ""
authors = ["Fatemeh Aghaie <fatemehaghaie2020math@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
django-countries = "^7.6.1"
coverage = "^7.4.4"
pillow = "^10.3.0"
python-decouple = "^3.8"
redis = "^5.0.4"
djangorestframework = "^3.15.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.7"
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
